move stream basic types from v1 to v2
