+++ src/stream/src/executor_v2/aggregation/mod.rs
@@ -36,11 +36,10 @@ use risingwave_storage::{Keyspace, StateStore};
 pub use row_count::*;
 use static_assertions::const_assert_eq;
 
-use crate::executor::PkDataTypes;
 use crate::executor_v2::aggregation::single_value::StreamingSingleValueAgg;
 use crate::executor_v2::error::{StreamExecutorError, StreamExecutorResult};
 use crate::executor_v2::managed_state::aggregation::ManagedStateImpl;
-use crate::executor_v2::Executor;
+use crate::executor_v2::{Executor, PkDataTypes};
 
 mod agg_call;
 mod agg_state;
