+++ src/stream/src/executor_v2/global_simple_agg.rs
@@ -23,8 +23,7 @@ use risingwave_common::catalog::Schema;
 use risingwave_common::error::Result;
 use risingwave_storage::{Keyspace, StateStore};
 
-use super::{Executor, ExecutorInfo, StreamExecutorResult};
-use crate::executor::{pk_input_array_refs, PkIndicesRef};
+use super::*;
 use crate::executor_v2::aggregation::{
     agg_input_array_refs, generate_agg_schema, generate_agg_state, AggCall, AggState,
 };
