+++ src/stream/src/executor_v2/managed_state/aggregation/extreme.rs
@@ -30,8 +30,8 @@ use risingwave_storage::{Keyspace, StateStore};
 
 use super::super::flush_status::BtreeMapFlushStatus as FlushStatus;
 use super::extreme_serializer::{variants, ExtremePk, ExtremeSerializer};
-use crate::executor::PkDataTypes;
 use crate::executor_v2::aggregation::{AggArgs, AggCall};
+use crate::executor_v2::PkDataTypes;
 
 pub type ManagedMinState<S, A> = GenericExtremeState<S, A, { variants::EXTREME_MIN }>;
 pub type ManagedMaxState<S, A> = GenericExtremeState<S, A, { variants::EXTREME_MAX }>;
