+++ src/stream/src/executor_v2/managed_state/aggregation/extreme_serializer.rs
@@ -20,7 +20,7 @@ use risingwave_common::types::{
 };
 use smallvec::SmallVec;
 
-use crate::executor::PkDataTypes;
+use crate::executor_v2::PkDataTypes;
 
 type ExtremePkItem = Datum;
