+++ src/stream/src/executor_v2/mod.rs
@@ -12,14 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod error;
+use std::collections::{HashMap, HashSet};
+use std::fmt::Debug;
+use std::pin::Pin;
+use std::sync::Arc;
 
+use enum_as_inner::EnumAsInner;
 use error::StreamExecutorResult;
 use futures::stream::BoxStream;
-pub use risingwave_common::array::StreamChunk;
+use futures::Stream;
+use risingwave_common::array::column::Column;
+use risingwave_common::array::{ArrayImpl, ArrayRef, DataChunk, StreamChunk};
+use risingwave_common::buffer::Bitmap;
 use risingwave_common::catalog::Schema;
+use risingwave_common::error::Result;
+use risingwave_common::types::DataType;
+use risingwave_pb::common::ActorInfo;
+use risingwave_pb::data::barrier::Mutation as ProstMutation;
+use risingwave_pb::data::stream_message::StreamMessage;
+use risingwave_pb::data::{
+    AddMutation, Barrier as ProstBarrier, DispatcherMutation, Epoch as ProstEpoch, NothingMutation,
+    StopMutation, StreamMessage as ProstStreamMessage, UpdateMutation,
+};
+use smallvec::SmallVec;
+use tracing::trace_span;
 
-pub use super::executor::{Barrier, Message, Mutation, PkIndices, PkIndicesRef};
+use crate::task::{ActorId, DispatcherId, ENABLE_BARRIER_AGGREGATION};
 
 mod actor;
 pub mod aggregation;
@@ -28,6 +46,7 @@ mod batch_query;
 mod chain;
 mod debug;
 pub mod dispatch;
+mod error;
 mod filter;
 mod global_simple_agg;
 mod hash_agg;
@@ -45,17 +64,21 @@ mod rearranged_chain;
 pub mod receiver;
 mod simple;
 mod source;
-#[cfg(test)]
-mod test_utils;
 mod top_n;
 mod top_n_appendonly;
 mod top_n_executor;
 mod union;
 
+#[cfg(test)]
+mod integration_tests;
+#[cfg(test)]
+mod test_utils;
+
 pub use actor::Actor;
 pub use batch_query::BatchQueryExecutor;
 pub use chain::ChainExecutor;
 pub use debug::DebugExecutor;
+pub use dispatch::DispatchExecutor;
 pub use filter::FilterExecutor;
 pub use global_simple_agg::SimpleAggExecutor;
 pub use hash_agg::HashAggExecutor;
@@ -68,7 +91,7 @@ pub use merge::MergeExecutor;
 pub use mview::*;
 pub use project::ProjectExecutor;
 pub use rearranged_chain::RearrangedChainExecutor;
-pub(crate) use simple::{SimpleExecutor, SimpleExecutorWrapper};
+use simple::{SimpleExecutor, SimpleExecutorWrapper};
 pub use source::*;
 pub use top_n::TopNExecutor;
 pub use top_n_appendonly::AppendOnlyTopNExecutor;
@@ -132,4 +155,310 @@ pub trait Executor: Send + 'static {
     {
         Box::new(self)
     }
+}
+
+pub const INVALID_EPOCH: u64 = 0;
+
+pub trait ExprFn = Fn(&DataChunk) -> Result<Bitmap> + Send + Sync + 'static;
+
+/// Boxed stream of [`StreamMessage`].
+pub type BoxedExecutorV1Stream = Pin<Box<dyn Stream<Item = Result<Message>> + Send>>;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum Mutation {
+    Stop(HashSet<ActorId>),
+    UpdateOutputs(HashMap<(ActorId, DispatcherId), Vec<ActorInfo>>),
+    AddOutput(HashMap<(ActorId, DispatcherId), Vec<ActorInfo>>),
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct Epoch {
+    pub curr: u64,
+    pub prev: u64,
+}
+
+impl Epoch {
+    pub fn new(curr: u64, prev: u64) -> Self {
+        assert!(curr > prev);
+        Self { curr, prev }
+    }
+
+    pub fn inc(&self) -> Self {
+        Self {
+            curr: self.curr + 1,
+            prev: self.prev + 1,
+        }
+    }
+
+    pub fn new_test_epoch(curr: u64) -> Self {
+        assert!(curr > 0);
+        Self::new(curr, curr - 1)
+    }
+}
+
+impl Default for Epoch {
+    fn default() -> Self {
+        Self {
+            curr: 1,
+            prev: INVALID_EPOCH,
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct Barrier {
+    pub epoch: Epoch,
+    pub mutation: Option<Arc<Mutation>>,
+    pub span: tracing::Span,
+}
+
+impl Default for Barrier {
+    fn default() -> Self {
+        Self {
+            span: tracing::Span::none(),
+            epoch: Epoch::default(),
+            mutation: None,
+        }
+    }
+}
+
+impl Barrier {
+    /// Create a plain barrier.
+    pub fn new_test_barrier(epoch: u64) -> Self {
+        Self {
+            epoch: Epoch::new_test_epoch(epoch),
+            ..Default::default()
+        }
+    }
+
+    #[must_use]
+    pub fn with_mutation(self, mutation: Mutation) -> Self {
+        Self {
+            mutation: Some(Arc::new(mutation)),
+            ..self
+        }
+    }
+
+    #[must_use]
+    pub fn with_stop(self) -> Self {
+        self.with_mutation(Mutation::Stop(HashSet::default()))
+    }
+
+    // TODO: The barrier should always contain trace info after we migrated barrier generation to
+    // meta service.
+    #[must_use]
+    pub fn with_span(self, span: tracing::span::Span) -> Self {
+        Self { span, ..self }
+    }
+
+    pub fn is_to_stop_actor(&self, actor_id: ActorId) -> bool {
+        matches!(self.mutation.as_deref(), Some(Mutation::Stop(actors)) if actors.contains(&actor_id))
+    }
+
+    pub fn is_to_add_output(&self, actor_id: ActorId) -> bool {
+        matches!(
+            self.mutation.as_deref(),
+            Some(Mutation::AddOutput(map)) if map
+                .values()
+                .flatten()
+                .any(|info| info.actor_id == actor_id)
+        )
+    }
+}
+
+impl PartialEq for Barrier {
+    fn eq(&self, other: &Self) -> bool {
+        self.epoch == other.epoch && self.mutation == other.mutation
+    }
+}
+
+impl Mutation {
+    /// Return true if the mutation is stop.
+    ///
+    /// Note that this does not mean we will stop the current actor.
+    #[cfg(test)]
+    pub fn is_stop(&self) -> bool {
+        matches!(self, Mutation::Stop(_))
+    }
+}
+
+impl Barrier {
+    pub fn to_protobuf(&self) -> ProstBarrier {
+        let Barrier {
+            epoch, mutation, ..
+        }: Barrier = self.clone();
+        ProstBarrier {
+            epoch: Some(ProstEpoch {
+                curr: epoch.curr,
+                prev: epoch.prev,
+            }),
+            mutation: match mutation.as_deref() {
+                None => Some(ProstMutation::Nothing(NothingMutation {})),
+                Some(Mutation::Stop(actors)) => Some(ProstMutation::Stop(StopMutation {
+                    actors: actors.iter().cloned().collect::<Vec<_>>(),
+                })),
+                Some(Mutation::UpdateOutputs(updates)) => {
+                    Some(ProstMutation::Update(UpdateMutation {
+                        mutations: updates
+                            .iter()
+                            .map(|(&(actor_id, dispatcher_id), actors)| DispatcherMutation {
+                                actor_id,
+                                dispatcher_id,
+                                info: actors.clone(),
+                            })
+                            .collect(),
+                    }))
+                }
+                Some(Mutation::AddOutput(adds)) => Some(ProstMutation::Add(AddMutation {
+                    mutations: adds
+                        .iter()
+                        .map(|(&(actor_id, dispatcher_id), actors)| DispatcherMutation {
+                            actor_id,
+                            dispatcher_id,
+                            info: actors.clone(),
+                        })
+                        .collect(),
+                })),
+            },
+            span: vec![],
+        }
+    }
+
+    pub fn from_protobuf(prost: &ProstBarrier) -> Result<Self> {
+        let mutation = match prost.get_mutation()? {
+            ProstMutation::Nothing(_) => (None),
+            ProstMutation::Stop(stop) => {
+                Some(Mutation::Stop(HashSet::from_iter(stop.get_actors().clone())).into())
+            }
+            ProstMutation::Update(update) => Some(
+                Mutation::UpdateOutputs(
+                    update
+                        .mutations
+                        .iter()
+                        .map(|mutation| {
+                            (
+                                (mutation.actor_id, mutation.dispatcher_id),
+                                mutation.get_info().clone(),
+                            )
+                        })
+                        .collect::<HashMap<(ActorId, DispatcherId), Vec<ActorInfo>>>(),
+                )
+                .into(),
+            ),
+            ProstMutation::Add(adds) => Some(
+                Mutation::AddOutput(
+                    adds.mutations
+                        .iter()
+                        .map(|mutation| {
+                            (
+                                (mutation.actor_id, mutation.dispatcher_id),
+                                mutation.get_info().clone(),
+                            )
+                        })
+                        .collect::<HashMap<(ActorId, DispatcherId), Vec<ActorInfo>>>(),
+                )
+                .into(),
+            ),
+        };
+        let epoch = prost.get_epoch().unwrap();
+        Ok(Barrier {
+            span: if ENABLE_BARRIER_AGGREGATION {
+                trace_span!("barrier", epoch = ?epoch, mutation = ?mutation)
+            } else {
+                tracing::Span::none()
+            },
+            epoch: Epoch::new(epoch.curr, epoch.prev),
+            mutation,
+        })
+    }
+}
+
+#[derive(Debug, EnumAsInner, PartialEq)]
+pub enum Message {
+    Chunk(StreamChunk),
+    Barrier(Barrier),
+}
+
+impl<'a> TryFrom<&'a Message> for &'a Barrier {
+    type Error = ();
+
+    fn try_from(m: &'a Message) -> std::result::Result<Self, Self::Error> {
+        match m {
+            Message::Chunk(_) => Err(()),
+            Message::Barrier(b) => Ok(b),
+        }
+    }
+}
+
+impl Message {
+    /// Return true if the message is a stop barrier, meaning the stream
+    /// will not continue, false otherwise.
+    ///
+    /// Note that this does not mean we will stop the current actor.
+    #[cfg(test)]
+    pub fn is_stop(&self) -> bool {
+        matches!(
+            self,
+            Message::Barrier(Barrier {
+                mutation,
+                ..
+            }) if mutation.as_ref().unwrap().is_stop()
+        )
+    }
+
+    pub fn to_protobuf(&self) -> Result<ProstStreamMessage> {
+        let prost = match self {
+            Self::Chunk(stream_chunk) => {
+                let prost_stream_chunk = stream_chunk.to_protobuf();
+                StreamMessage::StreamChunk(prost_stream_chunk)
+            }
+            Self::Barrier(barrier) => StreamMessage::Barrier(barrier.clone().to_protobuf()),
+        };
+        let prost_stream_msg = ProstStreamMessage {
+            stream_message: Some(prost),
+        };
+        Ok(prost_stream_msg)
+    }
+
+    pub fn from_protobuf(prost: &ProstStreamMessage) -> Result<Self> {
+        let res = match prost.get_stream_message()? {
+            StreamMessage::StreamChunk(ref stream_chunk) => {
+                Message::Chunk(StreamChunk::from_protobuf(stream_chunk)?)
+            }
+            StreamMessage::Barrier(ref barrier) => {
+                Message::Barrier(Barrier::from_protobuf(barrier)?)
+            }
+        };
+        Ok(res)
+    }
+}
+
+pub type PkIndices = Vec<usize>;
+pub type PkIndicesRef<'a> = &'a [usize];
+pub type PkDataTypes = SmallVec<[DataType; 1]>;
+
+/// Get clones of inputs by given `pk_indices` from `columns`.
+pub fn pk_input_arrays(pk_indices: PkIndicesRef, columns: &[Column]) -> Vec<ArrayRef> {
+    pk_indices
+        .iter()
+        .map(|pk_idx| columns[*pk_idx].array())
+        .collect()
+}
+
+/// Get references to inputs by given `pk_indices` from `columns`.
+pub fn pk_input_array_refs<'a>(
+    pk_indices: PkIndicesRef,
+    columns: &'a [Column],
+) -> Vec<&'a ArrayImpl> {
+    pk_indices
+        .iter()
+        .map(|pk_idx| columns[*pk_idx].array_ref())
+        .collect()
+}
+
+/// `StreamConsumer` is the last step in an actor.
+pub trait StreamConsumer: Send + 'static {
+    type BarrierStream: Stream<Item = Result<Barrier>> + Send;
+
+    fn execute(self: Box<Self>) -> Self::BarrierStream;
 }
