+++ src/stream/src/executor_v2/source.rs
@@ -33,8 +33,8 @@ use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver};
 
 use super::error::StreamExecutorError;
 use super::monitor::StreamingMetrics;
-use super::{BoxedExecutor, BoxedMessageStream, Executor};
-use crate::executor::{Barrier, ExecutorBuilder, Message, PkIndices, PkIndicesRef};
+use super::*;
+use crate::executor::ExecutorBuilder;
 use crate::task::{ExecutorParams, LocalStreamManagerCore};
 
 /// [`SourceExecutor`] is a streaming source, from risingwave's batch table, or external systems
@@ -322,7 +322,6 @@ mod tests {
     use tokio::sync::mpsc::unbounded_channel;
 
     use super::*;
-    use crate::executor::{Barrier, Epoch};
 
     #[tokio::test]
     async fn test_table_source() -> Result<()> {
